requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
charset-normalizer==3.2.0
frozenlist==1.4.0
greenlet==2.0.2
idna==3.4
multidict==6.0.4
psycopg2==2.9.6
py-cord==2.4.1
python-dotenv==1.0.0
SQLAlchemy==2.0.18
typing_extensions==4.7.1
yarl==1.9.2
